+++ NEWS.md
@@ -1,9 +1,10 @@
 centerline 0.0.2 (2024-05-07)
 =========================
 
-### BUG FIXES
-
-  * Now `cnt_path` family returns a one `LINESTRING`, not a collection of `LINESTRINGS`
+### UPDATES
+  
+  * `cnt_path_guess` performs 1.25 times faster due to a wiser usage of the `igraph` and `geos` packages. The `tidygraph` dependency has been removed.
+  * Now, the `cnt_path` family returns one `LINESTRING`, not a collection of `LINESTRINGS`.
 
 centerline 0.0.1 (2024-03-16)
 =========================